+++ CHANGES.md
@@ -8,6 +8,7 @@ Changes
     - IMPROVEMENT: improve FPS calculation in browsers
     - IMPROVEMENT: parallelize start/reload/stop of all browsers
     - IMPROVEMENT: better logging by using mainly IPC mechanism and catching also worker console
+    - IMPROVEMENT: fix preview better into the canvas
     - CLEANUP: use more consistent debug logging
 
 - 1.2.0